moved some files added store implemented first reducer for userdata
