+++ App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AppContainer from './app/components/CustomNavigation/Navigation';
-import store from './app/store/index';
+import store from './app/store/Index';
 import { Provider } from 'react-redux';
 
 export default class App extends React.Component {
