+++ app/reducers/Index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import Userdata from './Userdata';
+
+export default combineReducers({
+    Userdata
+});
